[Android] Always set a non-null Device.Info
When we run under the context of layoutlib the `Context` object
that is created will not implement IDeviceInfoProvider. All this
means is that we will not get change notifications when the
orientation changes, but that's ok! This won't happen anyway.

If we instantiate the Device.Info object then everywhere else in
the code will be able to get access to the screen properties.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=44893